+++ scripts/add_electricity.py
@@ -419,6 +419,7 @@ def attach_wind_and_solar(
     line_length_factor: float = 1.0,
     landfall_lengths: dict = None,
     trajectories: pd.DataFrame = pd.DataFrame(),
+    planning_horizon: int = None,
 ) -> None:
     """
     Attach wind and solar generators to the network.
@@ -444,6 +445,9 @@ def attach_wind_and_solar(
     trajectories : pd.DataFrame, optional, by default None
         DataFrame containing the trajectories for the current pyear to attach (p_nom_min and p_nom_max). When
         provided, these values override any p_nom_max defined in the profile itself.
+    planning_horizon: int, optional
+        Planning horizon for which renewable profiles should be added. Optional input defaults to None. If
+        year is part of the coordinates and planning_horizon is None, then the first year is used by default.
     """
     add_missing_carriers(n, carriers)
 
@@ -460,8 +464,10 @@ def attach_wind_and_solar(
             if ds.indexes["bus"].empty:
                 continue
 
-            # if-statement for compatibility with old profiles
-            if "year" in ds.indexes:
+            # if-statement for compatibility with old profiles and with PECD data
+            if planning_horizon is not None:
+                ds = ds.sel(year=planning_horizon, drop=True)
+            elif "year" in ds.indexes:
                 ds = ds.sel(year=ds.year.min(), drop=True)
 
             ds = ds.stack(bus_bin=["bus", "bin"])